+++ src/inst.h
@@ -1,7 +1,7 @@
 #ifndef INST_H
 #define INST_H
 
-#include <stddef.h>
+#include <stdint.h>
 #include "mat4.h"
 #include "mat.h"
 
@@ -12,12 +12,12 @@ typedef struct inst {
   mat4      transform;
   mat4      inv_transform;
   vec3      min;
-  size_t    id;   // (mat type << 28) | (mat id << 16) | (inst id & 0xffff)
+  uint32_t  id;   // (mat type << 28) | (mat id << 16) | (inst id & 0xffff)
   vec3      max;
-  size_t    ofs;  // Offset into tris/trisData/indices/bvh_nodes
+  uint32_t  ofs;  // Offset into tris/trisData/indices/bvh_nodes
 } inst;
 
-void inst_create(inst *inst, size_t inst_idx, const mat4 transform,
-    const mesh *mesh, const bvh *bvh, size_t mat_type, const mat *mat);
+void inst_create(inst *inst, uint32_t inst_idx, const mat4 transform,
+    const mesh *mesh, const bvh *bvh, uint32_t mat_type, const mat *mat);
 
 #endif
