+++ src/mesh.h
@@ -1,19 +1,19 @@
 #ifndef MESH_H
 #define MESH_H
 
-#include <stddef.h>
+#include <stdint.h>
 
 typedef struct vec3 vec3;
 typedef struct tri tri;
 typedef struct tri_data tri_data;
 
 typedef struct mesh {
-  size_t    tri_cnt;
+  uint32_t  tri_cnt;
   tri       *tris;
   tri_data  *tris_data;
 } mesh;
 
-void mesh_init(mesh *m, size_t tri_cnt);
-void mesh_load_obj(mesh *m, const char *path, size_t tri_cnt, size_t vertex_cnt, size_t normal_cnt, size_t uv_cnt);
+void mesh_init(mesh *m, uint32_t tri_cnt);
+void mesh_load_obj(mesh *m, const char *path, uint32_t tri_cnt, uint32_t vertex_cnt, uint32_t normal_cnt, uint32_t uv_cnt);
 
 #endif
