+++ src/scene.h
@@ -1,7 +1,7 @@
 #ifndef SCENE_H
 #define SCENE_H
 
-#include <stddef.h>
+#include <stdint.h>
 #include "view.h"
 #include "cam.h"
 
@@ -12,11 +12,11 @@ typedef struct bvh bvh;
 typedef struct tlas_node tlas_node;
 
 typedef struct scene {
-  size_t      mesh_cnt;
+  uint32_t    mesh_cnt;
   mesh        *meshes;
-  size_t      inst_cnt;
+  uint32_t    inst_cnt;
   inst        *instances;
-  size_t      mat_cnt;
+  uint32_t    mat_cnt;
   mat         *materials;
   bvh         *bvhs;
   tlas_node   *tlas_nodes;
@@ -24,7 +24,7 @@ typedef struct scene {
   cam         cam;
 } scene;
 
-void scene_init(scene *scene, size_t mesh_cnt, size_t inst_cnt, size_t mat_cnt);
+void scene_init(scene *scene, uint32_t mesh_cnt, uint32_t inst_cnt, uint32_t mat_cnt);
 void scene_release();
 
 #endif
