+++ src/tlas.h
@@ -1,18 +1,18 @@
 #ifndef TLAS_H
 #define TLAS_H
 
-#include <stddef.h>
+#include <stdint.h>
 #include "vec3.h"
 
 typedef struct inst inst;
 
 typedef struct tlas_node {
   vec3      min;
-  size_t    children; // 2x 16 bit
+  uint32_t  children; // 2x 16 bit
   vec3      max;
-  size_t    inst;     // Leaf nodes only
+  uint32_t  inst;     // Leaf nodes only
 } tlas_node;
 
-void tlas_build(tlas_node *nodes, const inst *instances, size_t inst_cnt);
+void tlas_build(tlas_node *nodes, const inst *instances, uint32_t inst_cnt);
 
 #endif
